vim: update simple-ag
`:command -range` doesn't work if you specify a range greater than the
lines in the current file. `:command -count` is inappropriate here as
that allows a number before OR after command. Searching for numbers like
`:Ag 12345` is often useful, so that is not ideal.

Instead, use `:command -bang` to allow `:Ag!` to search all results.